requirements.txt
dateparser==1.1.8
django>=5.0.1
django-filter>=23.5
djangorestframework>=3.14.0
markdown>=3.5.2
requests>=2.31.0